Move WeightInfo into MultiSig pallet.
