+++ pallets/common/src/lib.rs
@@ -16,6 +16,4 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod traits;
-pub use traits::{
-    asset, balances, compliance_manager, group, multisig, nft, portfolio, CommonConfig,
-};
+pub use traits::{asset, balances, compliance_manager, group, nft, portfolio, CommonConfig};
