+++ pallets/common/src/traits/mod.rs
@@ -27,7 +27,6 @@ pub mod balances;
 pub mod checkpoint;
 pub mod compliance_manager;
 pub mod group;
-pub mod multisig;
 pub mod nft;
 pub mod portfolio;
 pub mod settlement;
